// web/apps/storage/src/api/schema/models/objectVersionItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Storage API
 * Management API for buckets, access keys, policies and the file browser
 * OpenAPI spec version: 1.0.0
 */
import type { ObjectVersionItemStorageClass } from './objectVersionItemStorageClass';

export interface ObjectVersionItem {
  versionId: string;
  isLatest: boolean;
  isDeleteMarker: boolean;
  size: number;
  etag: string;
  storageClass: ObjectVersionItemStorageClass;
  createdAt: string;
}